Rewrite of Lagrangian post-processing output. The particles postprocessing output is now based on the standard FVM mechanism, whose handling was extended for these purposes. There is no restriction on output in parallel mode anymore. Thanks to Cyril Baudry for updating the GUI.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@4979 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ src/mesh/cs_join_merge.c
@@ -3207,8 +3207,9 @@ cs_join_create_new_vertices(int                     verbosity,
       bft_error(__FILE__, __LINE__, 0,
                 _("  Inconsistent value found in cs_join_vertex_t struct.:\n"
                   "    Vertex %d is defined by:\n"
-                  "      %u - [%7.4le, %7.4le, %7.4le] - %lg\n"),
-                i, vtx.gnum, vtx.coord[0], vtx.coord[1], vtx.coord[2],
+                  "      %llu - [%7.4le, %7.4le, %7.4le] - %lg\n"),
+                i, (unsigned long long)vtx.gnum,
+                vtx.coord[0], vtx.coord[1], vtx.coord[2],
                 vtx.tolerance);
 
   } /* End of loop on vertices */
